% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIDmixture.R
\name{ColourFile}
\alias{ColourFile}
\title{Creates and modifies the colour file}
\usage{
ColourFile(colourADMX, Kseq, palette)
}
\arguments{
\item{colourADMX}{file name for the colour file}

\item{Kseq}{Vector of K to be plotted}

\item{palette}{Colour palette}
}
\description{
Creates and modifies the colour file
}
\examples{
\dontrun{Admixture_ModPlot("Q_file", "MAP_file.map", "Sort_file.txt", 2:2)}
}
\keyword{Admixture_colour}
