% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIDmixture.R
\name{plotAdmixture_MB2}
\alias{plotAdmixture_MB2}
\title{Funtion to generate the Admixture PLot}
\usage{
plotAdmixture_MB2(Q_is, Fam_is, Sort_is = "", Col_is = NULL,
  Kseq = 2:4, toPDF = F, colorPal = NULL, margin = 0.05,
  modding = F, KtoMod = 0)
}
\arguments{
\item{Q_is}{.Q file stem}

\item{Fam_is}{fam file}

\item{Sort_is}{Order in which the populations should be plotted}

\item{Col_is}{Filename for the Colour file}

\item{Kseq}{Vector of K to be plotted}

\item{toPDF}{If TRUE sends the plot to pdf}

\item{colorPal}{The colour palette that will be used to plot}

\item{margin}{defines marigns in plot}

\item{modding}{If TRUE enters plot modification}

\item{KtoMod}{Integer: selects the K to modfy}
}
\description{
Funtion to generate the Admixture PLot
}
\examples{
\dontrun{Admixture_ModPlot("Q_file", "MAP_file.map", "Sort_file.txt", 2:2)}
}
\keyword{Admixture_plot}
