% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIDmixture.R
\name{Admixture_ModPlot}
\alias{Admixture_ModPlot}
\title{Main funtion to generate and modify the Admixture PLot}
\usage{
Admixture_ModPlot(Q_file, fam_file = NULL, Sort_file = NULL,
  Kseq = 2:2, ColourPalette = NULL, mod_plot = F, KtoMod = 3,
  ToPDF = F)
}
\arguments{
\item{Q_file}{.Q file stem}

\item{fam_file}{.fam file}

\item{Sort_file}{Order in which the populations should be plotted}

\item{Kseq}{Vector of K to be plotted}

\item{ColourPalette}{A colour palette}

\item{mod_plot}{Logical. TRUE to}

\item{KtoMod}{= 3,}

\item{ToPDF}{= F}
}
\description{
Main funtion to generate and modify the Admixture PLot
}
\examples{
\dontrun{Admixture_ModPlot("Q_file", "MAP_file.map", "Sort_file.txt", 2:2)}
}
\keyword{Admixture_mod}
